services/orders/requirements.txt
flask>=2.0
requests
psycopg2-binary
prometheus-flask-exporter
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
